vcf2parquet::columndata
Enum ColumnData

pub enum ColumnData {
    Bool(BooleanBuilder),
    Int(Int32Builder),
    Float(Float32Builder),
    String(StringBuilder),
    ListInt(ListBuilder<Int32Builder>),
    ListFloat(ListBuilder<Float32Builder>),
    ListString(ListBuilder<StringBuilder>),
}

Stores arrow array builders for each column datatype

Variants

Bool(BooleanBuilder)

Boolean column

Int(Int32Builder)

Int32 column

Float(Float32Builder)

Float32 column

String(StringBuilder)

String column

ListInt(ListBuilder<Int32Builder>)

List of int32 column

ListFloat(ListBuilder<Float32Builder>)

List of float32 column

ListString(ListBuilder<StringBuilder>)

List of string column

Implementations

impl ColumnData

pub fn new( arrow_type: &DataType, length: usize, field_name: &str, nullable: bool, ) -> Self

Creates a new ColumnData based on arrow type, length and field name

pub fn push_null(&mut self)

Add a Null value in array

pub fn len(&self) -> usize

Get the length of internal array

pub fn is_empty(&self) -> bool

Check if array is empty (not used for now)

pub fn push_bool(&mut self, value: bool)

Add a boolean value in array, if it’s not a boolean array failled

pub fn push_i32(&mut self, value: Option<i32>)

Add a i32 value in array, if it’s not a integer array failled

pub fn push_f32(&mut self, value: Option<f32>)

Add a f32 value in array, if it’s not a float array failled

pub fn push_string(&mut self, value: String)

Add a string value in array, if it’s not a string array failled

pub fn push_veci32(&mut self, value: Vec<Option<i32>>) -> Result<()>

Add a vector of integer value in array, if it’s not a vector of integer array failled

pub fn push_vecf32(&mut self, value: Vec<Option<f32>>) -> Result<()>

Add a vector of float value in array, if it’s not a vector of float array failled

pub fn push_vecstring(&mut self, value: Vec<Option<String>>) -> Result<()>

Add a vector of string value in array, if it’s not a vector of string array failled

pub fn into_arc(self) -> Arc<dyn Array>

Convert ColumnData in Arrow2 array